requirements_MAE.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=1.4.0=pypi_0
appdirs=1.4.4=pyhd3eb1b0_0
astunparse=1.6.3=pypi_0
blas=1.0=mkl
blosc=1.21.3=h6c2663c_0
brotli=1.0.9=h2bbff1b_7
brotli-bin=1.0.9=h2bbff1b_7
brotlipy=0.7.0=py39h2bbff1b_1003
bzip2=1.0.8=he774522_0
ca-certificates=2023.01.10=haa95532_0
cachetools=5.3.0=pypi_0
certifi=2023.5.7=py39haa95532_0
cffi=1.15.1=py39h2bbff1b_3
cfitsio=3.470=h2bbff1b_7
charls=2.2.0=h6c2663c_0
charset-normalizer=3.1.0=pypi_0
click=8.0.4=py39haa95532_0
cloudpickle=2.2.1=py39haa95532_0
colorama=0.4.6=py39haa95532_0
contourpy=1.0.7=pypi_0
cryptography=39.0.1=py39h21b164f_0
cudatoolkit=11.3.1=h280eb24_10
cycler=0.11.0=pypi_0
cytoolz=0.12.0=py39h2bbff1b_0
dask-core=2023.4.1=py39haa95532_0
einops=0.6.1=pypi_0
flatbuffers=23.5.8=pypi_0
fonttools=4.39.3=pypi_0
freetype=2.12.1=ha860e81_0
fsspec=2023.4.0=py39haa95532_0
gast=0.4.0=pypi_0
giflib=5.2.1=h8cc25b3_3
google-auth=2.18.0=pypi_0
google-auth-oauthlib=1.0.0=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.54.0=pypi_0
h5py=3.8.0=pypi_0
icc_rt=2022.1.0=h6049295_2
idna=3.4=py39haa95532_0
imagecodecs=2021.8.26=py39hc0a7faf_1
imageio=2.26.0=py39haa95532_0
importlib-metadata=6.6.0=pypi_0
importlib-resources=5.12.0=pypi_0
intel-openmp=2023.1.0=h57928b3_46319
jax=0.4.9=pypi_0
joblib=1.2.0=pypi_0
jpeg=9e=h8ffe710_2
keras=2.12.0=pypi_0
kiwisolver=1.4.4=pypi_0
lcms2=2.12=h83e58a3_0
lerc=3.0=h0e60522_0
libaec=1.0.4=h33f27b4_1
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libclang=16.0.0=pypi_0
libdeflate=1.8=h2bbff1b_5
libpng=1.6.39=h8cc25b3_0
libtiff=4.5.0=h8a3f274_0
libuv=1.44.2=h8ffe710_0
libwebp=1.2.4=hbc33d0d_1
libwebp-base=1.2.4=h8ffe710_0
libxcb=1.13=hcd874cb_1004
libzopfli=1.0.3=ha925a31_0
locket=1.0.0=py39haa95532_0
lz4-c=1.9.4=h2bbff1b_0
m2w64-gcc-libgfortran=5.3.0=6
m2w64-gcc-libs=5.3.0=7
m2w64-gcc-libs-core=5.3.0=7
m2w64-gmp=6.1.0=2
m2w64-libwinpthread-git=5.0.0.4634.697f757=2
markdown=3.4.3=pypi_0
markupsafe=2.1.2=pypi_0
matplotlib=3.7.1=pypi_0
mkl=2023.1.0=h8bd8f75_46356
mkl-include=2022.1.0=h6a75c08_874
mkl-service=2.4.0=py39h2bbff1b_1
mkl_fft=1.3.6=py39hf11a4ad_1
mkl_random=1.2.2=py39hf11a4ad_1
ml-dtypes=0.1.0=pypi_0
msys2-conda-epoch=20160418=1
networkx=2.8.4=py39haa95532_1
numpy=1.23.5=pypi_0
numpy-base=1.24.3=py39h8a87ada_1
oauthlib=3.2.2=pypi_0
opencv-python=4.7.0.72=pypi_0
openjpeg=2.4.0=h4fc8c34_0
openssl=1.1.1t=h2bbff1b_0
opt-einsum=3.3.0=pypi_0
packaging=23.1=pypi_0
partd=1.2.0=pyhd3eb1b0_1
pillow=9.5.0=pypi_0
pip=23.0.1=py39haa95532_0
pooch=1.4.0=pyhd3eb1b0_0
protobuf=4.23.0=pypi_0
pthread-stubs=0.4=hcd874cb_1001
pyasn1=0.5.0=pypi_0
pyasn1-modules=0.3.0=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=23.0.0=py39haa95532_0
pyparsing=3.0.9=pypi_0
pysocks=1.7.1=py39haa95532_0
python=3.9.16=h6244533_2
python-dateutil=2.8.2=pypi_0
python_abi=3.9=2_cp39
pytorch=1.10.1=py3.9_cuda11.3_cudnn8_0
pytorch-mutex=1.0=cuda
pywavelets=1.4.1=py39h2bbff1b_0
pyyaml=6.0=py39h2bbff1b_1
requests=2.30.0=pypi_0
requests-oauthlib=1.3.1=pypi_0
rsa=4.9=pypi_0
scikit-image=0.19.3=py39hd77b12b_1
scikit-learn=1.2.2=pypi_0
scipy=1.10.1=pypi_0
setuptools=66.0.0=py39haa95532_0
six=1.16.0=pypi_0
sklearn=0.0.post5=pypi_0
snappy=1.1.9=h6c2663c_0
sqlite=3.41.2=h2bbff1b_0
tbb=2021.5.0=h2d74725_1
tensorboard=2.12.3=pypi_0
tensorboard-data-server=0.7.0=pypi_0
tensorflow=2.12.0=pypi_0
tensorflow-estimator=2.12.0=pypi_0
tensorflow-intel=2.12.0=pypi_0
tensorflow-io-gcs-filesystem=0.31.0=pypi_0
termcolor=2.3.0=pypi_0
threadpoolctl=3.1.0=pypi_0
tifffile=2021.7.2=pyhd3eb1b0_2
timm=0.4.12=pypi_0
tk=8.6.12=h8ffe710_0
toolz=0.12.0=py39haa95532_0
torchaudio=0.10.1=py39_cu113
torchvision=0.11.2=py39_cu113
tqdm=4.65.0=pypi_0
typing_extensions=4.5.0=pyha770c72_0
tzdata=2023c=h04d1e81_0
urllib3=1.26.15=py39haa95532_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
werkzeug=2.3.4=pypi_0
wheel=0.38.4=py39haa95532_0
win_inet_pton=1.1.0=py39haa95532_0
wrapt=1.14.1=pypi_0
xorg-libxau=1.0.9=hcd874cb_0
xorg-libxdmcp=1.1.3=hcd874cb_0
xz=5.2.6=h8d14728_0
yaml=0.2.5=he774522_0
zfp=0.5.5=hd77b12b_6
zipp=3.15.0=pypi_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.2=h19a0ad4_0
